requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
tldextract>=3.4.0
rapidfuzz>=3.0.0
dnspython>=2.4.0
python-dateutil>=2.8.0
lxml>=4.9.0
